Changes tick & update rates to integers

+++ JankWorks.Game/source/Hosting/HostParameters.cs
@@ -3,7 +3,7 @@ namespace JankWorks.Game.Hosting
 {
     public struct HostParameters
     {
-        public float TickRate { get; set; }
+        public ushort TickRate { get; set; }
 
         public static HostParameters Default => new HostParameters()
         {
